third_party/blink/web_tests/svg/dynamic-updates/SVGRectElement-svgdom-width-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' property of the SVGRectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rectElement.width.baseVal.value is 50
PASS rectElement.width.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
